🐛(utils) ignore grep error if targeted commit does not exists
In the tag function we are looking for existing commit. To do this we
try to fount it in the log using the grep commend. This command exit and
return an error code if there is no match. In our case we just want to
ignore this case and continue the function.